:wrench: [#724] Disable vite public dir and make base relative
Vite copies the contents of public/\* as is into dist, which conflicts
with the storybook static build. Furthermore, for how we use and
distribute the SDK, the index.html and sdk.html are not relevant
anyway, they're only used during local development with CRA.

The base change from / to ./ should make things work better when
deployed/hosted on an arbitrary subpath.