:adhesive_bandage: [#724] Symlink fonts to root of project for vite
We cannot yet move the fonts to the public dir or change the SCSS that
loads the fonts, as that would break the actual CRA-based build which
now correctly ships the fonts.

Because of the relative URLs used in the font-face declaration, Vite
expects the fonts to exist in the root of the project.