:art: [#724] Address @charset warning emitted by vite-css
The formio styles include an invalid at-rule in the middle of the file,
which vite-css correctly warns about.

The workaround updates the SCSS to use @use rather than @import to
ensure the CSS is included rather than referenced, and this in turn
makes it compatible with the sass preprocessor options to disable the
charset at-rule, effectively getting rid of the warning.