:alien: [#724] Fix loading/resolving the font-awesome font files
CRA seems to properly translate the relative path and include the files
while Vite requires some extra hints to properly point to the package
location before it can find the font files.

Unfortunately, both setups are not compatible, so in the vite config
we override the sass-variable to pass down to the sass module import,
where the default is set to the current, legacy behaviour.