:iphone: [#724] Disable mobile-viewport test
When running tests using @storybook/test-runner, we need to apply
the same viewport to get the proper responsive elements as we
get in the real browser/storybook. However, we just ripped out
the viewport support because it's super flaky. So, we disable
these tests here, but they should still run on Chromatic so
we still get feedback there (provided we don't disable the
Chromatic snapshotting).