:alien: [#724] Use different Dutch-locale import for flatpickr
Vite/Rollup seem to be much more aggressive in handling imports with
side-effects, resulting in the Dutch locale not getting registered on
the flatpickr instance.

Changing the import to a named import and manually assigning it fixes
the issues in the static storybook build, while not breaking the CRA
build (verified locally by creating a production build and using that
with the real backend).