:alien: [#724] Map error classes rather than their constructor names
This is due to a difference between webpack and Vite/Rollup - CRA/wp
is configured to not mangle class names in the minification process,
while Vite/Rollup do mangle them. One option is to disable class name
mangling which would probably only have a minor effect on the bundle
sizes, however it's just technically more correct to point to the
class itself rather than the name of the class in the runtime time,
as this will be more robust for future build tooling changes.