:truck: [open-formulieren/open-forms#4929] Move CoSignOld component
Moved component into its own file so in the future we can more easily
remove it, and to declutter index.jsx from a package entrypoint.

Next, we can export the nested routes from the CoSign package without
adding even more weight/content to the file itself in terms of
implementation.