:white_check_mark: [open-formulieren/open-forms#4929] Add tests for top-level Cosign component (start)
The top-level component is responsible for some state management while
the child route ./start actually ensures the right component gets
rendered.

Setting up the tests like this gets the boilerplate out of the way and
allows us to test from the user perspective for the more complicated
stuff in the cosign component.