:ambulance: Fix code-transformation during bundling and minifying in external dependency
Names get mangled/minified during production bundling, which
affects how the js-interpreter package this['Interpreter']
gets transformed. This makes things harder to debug, since
the likely root cause is the UMD shim in combination with
strict mode leading to 'this' not pointing to the global
(the window object in the browser), but the rename/mangling
also doesn't help of course.

Resolved by doing a find-and-replace of the code during
bundling to make this an explicit line referring to the
window global, as we only use this in a browser context
anyway.

js-interpreter is forked by formio in their
@formio/protected-eval module, which avoids Formio breaking
on 'eval' calls due to our CSP requirements (and frankly,
eval is just... not okay).

I am not sure how important this is and how often JS gets
evaluated - most of the Formio features around that are
not exposed by us (I can't think of anything that we *do*
expose), but there may be some lodash templating stuff -
particularly with file uploads) that could break on just
not supporting any flavour of eval. The plugin/workaround
can be removed if we can drop formio.js as a dependency
and replace it with our own renderer.