:recycle: [open-formulieren/open-forms#4929] Extract the initial data reference implementation into hook
Rather than sprinkling the param processing through various parts of
the code base and repeat the implementation, provide a custom hook that
encapsulates the implementation details.

The hook reads the reference from the query params, and provides a
helper to mutate the reference in (derived) react houter URLs.

Note that the window.location.origin as base is just a workaround to be
able to use the `URL` constructor and object for query string
manipulation. The actual real window.location is irrelevant, as that
state is managed by react router. We only care for the path (or rather
the path segment) that may have some query params attached to it.