:sparkles: [open-formulieren/open-forms#2177] Allow manually deleting features
Due to some formio magic, the `resetValue` flag is needed when performing the `setValue` call. Without this flag it isn't allowed to set a field value to `null` or `undefined`.We also need to use the value `null`, instead of `undefined`. This because setting the value to `undefined` isn't reflected as a value change (what we need, to re-render the component).
This is described in node_modules/formiojs/components/_classes/component/Component.js:2524