:recycle: [open-formulieren/open-forms#4929] Take initial data reference from query instead of prop
The FormStart component now grabs the initial data reference directly
from the URL parameters rather than requiring the parent component needing
to do this and pass it down as a prop. This is a first step towards not
needing to specify props at all to FormStart so that it can be included
in the data router routes (statically).