:technologist: [open-formulieren/open-forms#4929] Improve debug information
The app debug state is now displayed on the session-expired page/route,
which can give some more insight on what happens with the session
reset etc.

Additionally, when the expiry date is reset it's set to null, and the
delta calculations are off/weird because of Javascript doing math
on 'null' objects. Now, we display a dash to indicate there's no
session expiry at all.

This was discovered during testing if certain code is still required
or not.