:white_check_mark: [open-formulieren/open-forms#4929] Clean up and update LoginOptions tests
Made the tests more robust by doing proper URL parsing instead of
string comparisons, as the order of query string parameters is not
important or deterministic.

The router wrapper is updated so we can pass the current location
without monkeypatching the global window object, which also cleans
up the test a bit more.

Finally, the onFormStart callback is updated since we don't pass an
event any longer - it is handled on the form submit action already.