:construction_worker: [#724] Specify coverage flags for storybook/vitest
Storybook and vitest cover different parts of the codebase, and
coverage checks fail sometimes because one is still running/not
uploaded yet while the other one is. Tracking the coverage of these
separately allows us to better judge which job produces which
coverage.

Storybook tests should focus on interaction and not much on coverage,
whereas vitest tests should ensure sufficient coverage is achieved.