:recycle: [open-formulieren/open-forms#4929] Refactor RequireSubmission wrapper to not require dynamic props
The submission (which is retrieved from the state and 'highly' dynamic
can now be taken from the submission context rather than from a prop,
which makes moving this wrapper/component to a statically defined route
element achievable.

Additionally, the pattern of taking a component type and its props to
render it dynamically has fallen a bit out of favour, so we now
encourage to specify the child element as a react element with the
necessary data. Other routes will be updated accordingly and then the
deprecated approach can be removed.