:recycle: [open-formulieren/open-forms#4929] Refactor the session timeout monitoring
This is kind of between refactoring and bugfix...

The session timeout would run even on pages where this is not strictly
necessary, such as the form start page, but more importantly, on the
submission payment and confirmation pages. If you wouldn't complete
any tasks within the session timeout, the SDK would automatically
expire your client side session because the backend session is
expired. However, on these pages, the backend session is no longer
required and has been terminated anyway when the submission was
completed.

The refactor bit is encapsulating the useTimeout hook directly into the
SessionExpiryModal, which allows us to use this component without
passing any additional props, making it suitable for declaration in
static react router routes - and that's what our goal is for this
branch/feature/ticket/PR.

Finally, for devs using an session expiry larger than ~25 days, an
infinite re-rendering loop was fixed, but that shouldn't affect any
real users.