:arrow_up: [open-formulieren/open-forms#4929] Bump testing-library packages
Upgraded the packages in an attempt to debug why a new test would keep
running forever without crashing/erroring out, seemingly being stuck
in an infinite loop.

I suspected something odd with timers/setTimeout going on and there
are some open issues in testing-library/vitest related to that, but
this was not the problem.

I've decided to keep the upgrades since things *appear* to work with
the new versions, but of course only CI will decide that for real.
If we can keep this 'free' upgrade, that's a nice accident, if it's
problematic, I revert it.