:bug: [open-formulieren/open-forms#4929] Fixed a test-specific bug in useScrollIntoView
jest-dom doesn't provide the scrollIntoView method on HTML elements for
it's JS API, which explains why this code/flow appears to work
properly when using a real browser and simulating testing flows
manually, while being problematic in tests.

The problem here was ultimately that no mock was set up for this
method, causing the useEffect hook to crash because undefined is not
a function that can be called. This problem is then compounded because
React will render the nearest ErrorBoundary, which we have define in
most of our routes (that are currently under testing), and these
error boundaries use the ErrorMessage component too, which also calls
useScrollIntoView, which also crashes and seemingly gets stuck in an
infinite render, causing Vitest to keep running forever without
erroring or even timing out (presumably because the JS thread is just
stuck in an infinite loop).

The fool-proof way to mitigate this is to nullish check the function
before calling it, so that future oversights (forgetting to mock) will
not cause us to spend hours of debugging again. For the tests, the
scrolling behaviour is not relevant anyway.

A test is added to establish the assumption that this callback doesn't
exist, which protects against accidentally setting up a mock in this
test file.