:white_check_mark: [open-formulieren/open-forms#4929] Add test for summary -> status -> error submission flow
There's now a test to lock in the behaviour after confirming the
submission, getting the status page with spinner and redirect back
to summary page if there's a background processing error.

Now we can refactor the error message state handling to get these
props out of the Summary component.

The fake timer situation is... complicated. See
https://github.com/testing-library/react-testing-library/issues/1197