:coffin: [open-formulieren/open-forms#4929] Delete dead code
The error handling of the post method is obsoleted because the apiCall
helper itself already calls throwForStatus which results in an
exception if/when response.ok is false, so this block of code will
never execute.

This also further simplifies the helper to directly return the status
URL, which is the only bit of relevant information returned by the
API endpoint.