:recycle: [open-formulieren/open-forms#4929] Move failure handling entirely into StatusUrlPoller
The StatusUrlPoller component is tightly coupled with the particular
submission status API endpoint anyway - it already checks specific
props to determine whether it's failing or succeeding. Since the
failure handlers are the same in both places, we can localize this
behaviour/the details in the component itself and only pass the
particular route it should redirect/navigate the user to which
is responsible for display the error message.

This simplifies quite a bit of code. Maybe at some point we can
also get rid of the success callback used by the appointments
flow.