:truck: [open-formulieren/open-forms#4929] Declare app routes
We now have a single entry point for the SDK routes
that can easily be re-used in tests, storybook and by
the actual application. If needed, specific submodule
routes can be loaded with their appropriate import
paths.

One advantage of this is that component files can
follow the best practice of only exporting the
component as default without also exporting
the routes, and the route definitions are no
longer in the 'components' folder, which is more
semantically correct.