:arrow_up: Upgrade to vitest 3
Upgraded from vitest 2 to vitest 3, so now we only depend
on one version of Vite in the whole project.

Additionally, rollup was explicitly added to the dev
dependencies so that the type imports are picked up by
TS - it is a direct dependency of Vite already so
nothing really changed except for TS being able to
resolve the types.