:zap: [#76] Restructure applyPrefix utility for better bundle splitting
Moving this away from the formio scope/namespace allows us to use it
in the rest of the non-formio related code without pulling in all of
lodash into the main bundle, allowing us to defer loading and have
a faster time-to-first-interact for end-users.

The utility itself is of questionnable importance, but that's a
refactor for another day.