:coffin: [#76] Delete more dead code
More dead code that relied on Formio utils which caused our
bundles to be larger than necessary. The backend has been
updated a long time ago to provide the summary data to
display, rather than having to do this client-side.

The benefit is that no longer load anything from formio
outside of the formio-init chunk, which should result
in better perceived performance for end users.