:bug: Fix router wrapper around error boundary
Not sure why exactly this is or isn't working properly with
the built in decorator, but I suspect it has something to
do with the error boundary of react router itself/the
decorator and the error state not being reset properly
when the story remounts.

The workaround just creates a new memory router instance
on every render of the story, discarding any router state
between every story render which seems to solve the
problem.