:sparkles: [open-formulieren/open-forms#4510] Display backend validation errors
I've opted to collect all the errors at the top and group them by step,
rather than trying to weave this in the summary table. User research
from Utrecht showed that this was the best way to show validation
errors.

Ideally, there would be some aria-describedby options, but that requires
a lot more work to link everything together.