:bug: Define vite build dir based on build version
For release builds we want to the assets to be placed in `/dist/[release number]`.

This logic was present in the previous build setup, with CRA. In the re-structure of the SDK project, #724, the build process of CRA was replaced with vite. But this logic rule was lost in the process