:wrench: [#445] Enable skipLibCheck, unfortunately
We get some conflicts from leaflet libraries and some @storybook/test shenanigans,
additionally there are some errors from react-use. It's not feasible to address
these right now since we're aiming to get a foot in the door and get
started with typing our own code.

The offending files:

Found 12 errors in 8 files.

Errors  Files
     1  node_modules/@open-formulieren/leaflet-tools/lib/rd.d.ts:8
     1  node_modules/@vitest/expect/dist/chai.d.cts:16
     3  node_modules/react-leaflet-draw/src/index.d.ts:4
     3  node_modules/react-leaflet/lib/LayersControl.d.ts:9
     1  node_modules/react-leaflet/lib/SVGOverlay.d.ts:11
     1  node_modules/react-use/lib/useEnsuredForwardedRef.d.ts:1
     1  node_modules/react-use/lib/usePermission.d.ts:10
     1  node_modules/vitest/node_modules/@vitest/expect/dist/chai.d.cts:16

Hopefully in the future we can re-enable this again.