:label: [#445] Convert the Body component and stories to TS
Turns out we need the babel presets after all for the storybook
processing :( I'm not sure why things work without in the renderer
repo, but we'll figure it out at some point.

This strongly types the Body component with its 'component' prop,
allowing strings of intrinsic elements to be passed. Support for
custom react components is dropped, as it doesn't appear to be
used anywhere so this simplifies things a bit.