:arrow_up: [#445] Upgrade to @utrecht/components v5
Adapted the few places where we directly import the mixin
for certain components, which has been moved into the package
src and is no longer in the dist/css directory.

Other imports have been updated to refer to the package itself,
as that resolves to the CSS file inside, and we can now use
the @use rule in a bunch of places rather than sticking with
the deprecated @import.