:coffin: [open-formulieren/formio-renderer#36] Remove obsoleted editgrid code
The editgrid code was ported from the SDK to the formio-renderer
package, so after replacing the usages here we can remove the code.

The CSS was also ported, except for the mobile center-content rule,
which has now been scoped to the Open Forms theme itself to not
invade on users who built their own design tokens and/or have mobile
rules taken into account already.