:bug: [open-formulieren/open-forms#5151] Fix empty value for map component
Newer versions were already patched in
6cfeb2958e63b0fb4c18d7becbff44642fbdd024.

This empty value is used when the map component is contained in a
fieldset that is (conditionally) hidden and the clearOnHide behaviour
is triggered. This resulted in a string being sent for the map
component value, while an array of latitude/longitude is expected,
and this causes the form step not to validate.

(cherry picked from commit 83576b7f95c32d701bef56f02dd6f30bd4c8dc5b)