:recycle: [open-formulieren/open-forms#4917] Rewrite form navigation styles
The markup and styles for the link/previous link are updated to still
enact the behaviour of the utrecht-link component with all the
desired interactive states (hover, focus, active...) while enabling
icons to be displayed with a column gap rather than enforcing margins.

The gap values are specified via design tokens, but we fall back to
existing NL Design System (button) tokens to hopefully make upgrading
less painful.

This particular approach also takes care that the icons in the links
don't get any text decoration while remaining clickable and part of
the link. This is quite tricky, because apparently browsers don't
render this underline on inline-block elements inside anchor tags,
and this worked as long as we applied block/inline styling on the
anchor tag itself - icons had no text decoration. However, when
switching to flex box, this automatically makes the child items behave
in block display mode, which causes the text decoration to be
rendered too, leading to ugly links.

This is solved by disabling the link styles at the anchor level (never
show any text decoration), and deferring those to the child element
that contains the link text by copying/extending some styles.