:recycle: [open-formulieren/open-forms#4917] Ensure that link styles with icons are properly applied
Golden tip from @Robbert to wrap the icons in utrecht-icon, which
ensures that the icon itself is again using 'display: inline-block'
causing the text decoration to not be rendered, while still being able
to use flexbox for the anchor tag itself.

There was also a bug where the `className` prop passed to the Link
component would completely overwrite the computed classname in our
Anchor component (building on top of utrecht-link), breaking hover
styles etc.