:recycle: [open-formulieren/open-forms#4917] Remove onPrevPage prop
All this prop seemed to do was programmatically call the navigate
hook, but this is exactly the same behaviour as just passing the 'to'
prop to the Link component, so it has been removed.

There's only one prop like it in a lower level component left becaues
it's actually used to do state updates in a parent component.