:label: [#445] Properly type check polymorphism for the Link/Anchor components
This was a bit tricky and we have to make a detour via a string union
to render the expected component (link vs link-button), as both
React components render an anchor under the hood so structurally they
are the same. This would be one of those few cases where you'd prefer
nominal typing, I guess...

The original 'bug' introduced in this branch didn't show because the
introduction page component was not TS yet, and the broken 'component'
prop was not picked up by the compiler. This has now been rectified.