:bug: [open-formulieren/open-forms#5242] Fixing browser translations crashing the SDK.
Browser translations, in participle the translation functions of Chrome, change the DOM tree. This causes React to lose track of DOM nodes, which leads to the SDK crashing.

With this we lose some accessibility, but this seems to be the only reliable solution that would "solve" the issue for any and all browsers and translation plugins.