:bug: [open-formulieren/open-forms#5242] Prevent browsers from translating the SDK.
Browser translations, in particular the translation functions of Chrome, change the DOM tree. This causes React to lose track of DOM nodes, which leads to the SDK crashing.

By disabling the browser (and possible plugin) translations, we can prevent these DOM changes. With this we do lose (quite) some accessibility, but this seems to be the only reliable solution that would "solve" the issue for any and all browsers and translation plugins.

Backport-Of: open-formulieren/open-forms#5242