:sparkles: [open-formulieren/open-forms#5191] Custom defined Leaflet map delete control
The default delete control expects you to use the delete toolbar for precise control over which features are deleted. Because we only allow 1 feature to be drawn at a time, the delete toolbar is a bit too precise and requires the enduser to have some Leaflet functionality knowledge to effectively work with the map component.

We can define a `onDeleteStart` eventHandler on the `EditControl`. This would allow us to delete the drawn features without using the toolbar, but if a user declines the new `window.confirm()` the toolbar menu cannot automatically be closed. To provide delete functionality completely without the toolbar, we need to define our own delete control.