:sparkles: [open-formulieren/open-forms#5191] Fit map to geometry bounds
This ensures that the map shows the entire geometry shape in the summary view. Additionally, this re-fits the map when the user draws a shape.

To prevent the map/zoom from jumping around too much, the `map.fitBounds()` is only done for polygon and linestring features. For "point" features the `map.fitBounds` doesn't add anything, seeing how a point cannot be outside the bounds. This also ensures that the previously know functionality and expectations of the map remain.