:bug: [open-formulieren/open-forms#5240] Fix calendar widget date value format
The default calendar widget value format is ISO 8601 datetime, but our
date component only requires the date part and not the time part. The
time part is actually causing problems when attempting to pause a form,
because the backend validation trips on the unexpected time information.

This has been a bug since Open Forms 1.0. At that time, the time info
was truncated from the value on submit of the form (so when you
navigate to the next step), which cleans up the data, but that doesn't
work/get called when you snapshot the form data for pausing it.

The fix is simple after reverse engineering how Formio gets the values
from its widgets and how/where they are instantiated, and then
figuring out which magical option you need to specify to change the
behaviour, but at least it's possible in a simple way.