:wrench: [#5296] Update config file for bumping version
With open-formulieren/open-forms#5201, we now use bump-my-version instead of bump2version. There
was some trouble with updating the `package.json` file: it was unable to find the current version.
According to the docs (https://callowayproject.github.io/bump-my-version/reference/configuration/#configuration-files),
a .cfg file is legacy, so I have replaced it with a .toml file based on the configuration we have
inside `pyproject.toml` for the backend. This fixed the issues with updating the `package.json` file