:bug: [open-formulieren/open-forms#5266] Ensure new initial data reference is kept when starting a form with an existing submission
When a user would start a submission with initial data reference A, cancel this submission - either
by closing the current window/tab or navigating away - and start a new submission (for the same form) with
initial data reference B, a prompt is shown whether to continue with the existing submission or cancel and start
over. Continuing with the existing submission would mean discarding the initial data reference B, and the user
was not aware of this. Also, cancelling the existing submission would cause initial data reference B to be lost.

We now combine the form uuid and initial data reference to determine if an existing submission is available.
This means a user will not be asked to continue a submission with initial data reference A if they start a
new submission with initial data reference B. We also ensure initial data reference B is kept when an existing
submission is cancelled.