:recycle: Switch to session storage
This will scope a submission within a single tab/window. Note that the `useSessionStorage` hook does not
return a hook to remove an entry from the storage, so we create our own that sets the submission ID of a
particular form to an empty string.

Also converted some of the files to typescript