:bug: [open-formulieren/open-forms#5266] Don't offer to re-use submission for (new) initial data reference
The presence of an initial data reference parameter implies it's a new
reference, which would invalidate any existing submissions stored in
the local storage. So, the easiest (but not most elegant) patch for the
observed problem is to ignore any existing submission ID from the
local storage (or session storage, if you will) when this query param
is present.

This works because the parameter is only used to bootstrap the process
and kept around long enough to be able to send it with the submission
create call. After that, it's no longer passed around in the URL, and
navigating back to the form start step while on step 2 or further
will present the 'continue or reset' screen as usual because the
parameter is missing.

When the parameter is present, it's because of the user clicking a link
or manually pasting the link back into the browser, at which point we
can reasonably assume they're trying to start a new form submission.

This patch does not interfere with hitting F5 anywhere in the form,
even refreshing on the start page again has the expected behaviour,
since a submission still needs to be created then.