:green_heart: Increase test timeouts
The lazy loading (and instrumenting?) of modules requires longer
timeouts on CI. The test failures were not reproducible on my
local machine, except when running the actual workflows with
act. I suspect the coverage option and associated code
instrumentation to lead to longer delays, and the test times
out before all code was processed and loaded.