:package: [#850] Provide explicit exports for our formio module and template library
The bundle splitting rework to optimize the chunks broke external
storybooks trying to integrate our Formio module and template library
because they're no longer part of the exports of the UMD bundle.

We can't make them part of that again, because that breaks lazy loading.

Instead, there are now explicit exports for people wishing to integrate
these formio components in their own storybook.

Backport-of: #851