Fixed the behaviour of `selected` property
What we had until now was that the selected property was always a
boolean. This meant that in order to distinguish a selected child in the
backend was really complicated (needed the component configuration and
the `enableSelection` value too). Now we simply assign null when the
`enableSelection` is False and a boolean when it's true.