:recycle: [#445] Refactory/convert GenericSummary to typescript
Applied some discriminated union semantics to the props to make it
clear when certain props are required and when they're optional,
which makes the call sites in some places simpler/more expressive.

This also factors out the SUBMISSION_ALLOWED constant, which has been
superseded by the union in the Submission type - the use of TS here
actually allows us to reduce runtime code and make use of 'magic
strings' in a type-safe manner.