:alien: [open-formulieren/open-forms#5133] The RenderSettingsProvider was renamed
Just fixing up the imports and names is sufficient. The
components prop is new and required, but for our current
usage we can just pass a static empty array.

It is defined at the module level as a constant so that
there are no new objects created at render time, which
would break the stable identity of the array instance and
cause unnecessary re-renders.