:heavy_plus_sign: [open-formulieren/open-forms#5133] Make fast-deep-equal an explicit dependency
We already include it through react-leaflet-draw, which is why this
package is picked over fast-equals which seems better maintained.

The fast-deep-equal comparison is necessary because we can't rely on
JSON.stringify for a stable key order when comparing different formio
configurations, and we only want to update state if there are actual
meaningful changes, otherwise we break stable identities which causes
unnecessary renders down the component tree.