:sparkles: [open-formulieren/open-forms#5133] Define submission step API/data interaction
The legacy form step component performs many ad-hoc calls and building
of API endpoints in the component itself, which can be abstracted in a
nicer way in our data layer so that we get more meaninful function
calls in the React code.

This also prepares a bunch of mocks to move into the api-mocks layer
rather than being duplicated, and we can make them type-safe(r) along
the way.